fix nsmeof packages task1416
Task: https://github.com/jdi-testing/jdi-light/issues/1416
