+++ jdi-light-bdd-no-po-tests/src/test/java/cucmbertests/Runner.java
@@ -1,4 +1,4 @@
-package cucmberTests;
+package cucmbertests;
 
 import com.epam.jdi.light.elements.composite.Form;
 import com.epam.jdi.light.elements.composite.WebPage;
