+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/composite/section/listgroup/DisabledItemsTests.java
@@ -1,4 +1,4 @@
-package io.github.epam.bootstrap.tests.composite.section.listGroup;
+package io.github.epam.bootstrap.tests.composite.section.listgroup;
 
 import io.github.epam.TestsInit;
 import org.testng.annotations.BeforeMethod;
@@ -6,7 +6,7 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import static io.github.com.StaticSite.bsPage;
-import static io.github.com.pages.BootstrapPage.listGroupFlush;
+import static io.github.com.pages.BootstrapPage.listGroupDisabledItems;
 import static io.github.epam.states.States.shouldBeLoggedIn;
 import static org.hamcrest.Matchers.is;
 
@@ -15,7 +15,7 @@ import static org.hamcrest.Matchers.is;
  * Email: delnote@gmail.com; Skype: Dmitrii Pavlov
  */
 
-public class FlushTests extends TestsInit {
+public class DisabledItemsTests extends TestsInit {
 
     private String text1 = "Cras justo odio";
     private String text2 = "Dapibus ac facilisis in";
@@ -38,27 +38,31 @@ public class FlushTests extends TestsInit {
     }
 
     @Test
-    public void initTests() {
-        listGroupFlush.listGroup.is().size(5);
-        listGroupFlush.is()
+    public void isValidationTests() {
+        listGroupDisabledItems.listGroup.is()
+                .size(5);
+        listGroupDisabledItems.is()
                 .displayed()
                 .enabled()
                 .core()
-                .hasClass("list-group list-group-flush");
+                .hasClass("list-group");
+        listGroupDisabledItems.listGroup.get(1).is()
+                .hasClass(listClass + " disabled")
+                .attr("aria-disabled", "true");
     }
 
     @Test(dataProvider = "listData")
-    public void listGroupTests(int num, String text) {
-        listGroupFlush.listGroup.get(num).is()
+    public void listGroupTextTests(int num, String text) {
+        listGroupDisabledItems.listGroup.get(num).is()
                 .text(text)
                 .css("font-size", is("14px"));
     }
 
-    @Test(dataProvider = "listData")
-    public void listGroupIsValidationTests(int num, String text) {
-        listGroupFlush.listGroup.get(num).is()
+    @Test()
+    public void listGroupClassesIsDisabledValidationTest() {
+        listGroupDisabledItems.listGroup.get(1).is()
                 .displayed()
-                .enabled()
+                .disabled()
                 .core()
                 .hasClass(listClass);
     }
