+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/composite/section/listgroup/Horizontal.java
@@ -1,4 +1,4 @@
-package io.github.epam.bootstrap.tests.composite.section.listGroup;
+package io.github.epam.bootstrap.tests.composite.section.listgroup;
 
 import io.github.epam.TestsInit;
 import org.testng.annotations.BeforeMethod;
